Removes forms, adds enemy, bullet, health
Removes the forms from the levels (still needs to clean a lot of HUD stuff from the old project and the main menu). Adds bullets for the player to attack (with sound). Adds a simple enemy (placeholder sprite) which chases the player and dies when hit with enough bullets. Player now has health and can die (on debug message only, still needs the whole programming part).