Adds colorblind, photophobia, towerEnemy. Enemies now Spawn per room.
Added colorblind effects and photophobia (bloom). Now a manager randomly selects an impairment every new game and remove the others accordingly. Tower Enemy created, animated and is working normally. It shoots bullets that damage the player. Placeholder sprite substituted for slime's. Player's bullets now spawn in the right region of the player's eye, but the sprite itself does not rotate (yet). Dyslexic font added in the project. Will be used for the HUD and UI later. New sprites for the rooms' tiles added.