Fix player animation, enemy spawning and camera follow
Now player animation will follow first the shooting direction, then the movement direction, if not shooting. When not walking, it will face the shooting direction.
The enemies now spawn in a random amount for each room. The amount is predetermined during the creation of the dungeon. If all enemies in the room are destroyed they won't respawn. Else, the remaining are deleted and respawned when the player re-enters the room.
Camera follows the player faster now. It was very slow.